Tidy fixes
